+++ eras/conway/impl/src/Cardano/Ledger/Conway/PParams.hs
@@ -37,7 +37,7 @@ module Cardano.Ledger.Conway.PParams (
   ppDRepVotingThresholdsL,
   ppMinCommitteeSizeL,
   ppCommitteeTermLimitL,
-  ppGovActionExpirationL,
+  ppGovActionLifetimeL,
   ppGovActionDepositL,
   ppDRepDepositL,
   ppDRepActivityL,
@@ -45,7 +45,7 @@ module Cardano.Ledger.Conway.PParams (
   ppuDRepVotingThresholdsL,
   ppuMinCommitteeSizeL,
   ppuCommitteeTermLimitL,
-  ppuGovActionExpirationL,
+  ppuGovActionLifetimeL,
   ppuGovActionDepositL,
   ppuDRepDepositL,
   ppuDRepActivityL,
@@ -88,7 +88,7 @@ import Numeric.Natural (Natural)
 -- * @dRepVotingThresholds@
 -- * @minCommitteeSize@
 -- * @committeeTermLimit@
--- * @govActionExpiration@
+-- * @govActionLifetime@
 -- * @govActionDeposit@
 -- * @dRepDeposit@
 -- * @dRepActivity@
@@ -150,8 +150,8 @@ data ConwayPParams f era = ConwayPParams
   , cppCommitteeTermLimit :: !(HKD f Natural) -- TODO: This too should be EpochNo
 
   -- ^ The Constitutional Committee Term limit in number of Slots
-  , cppGovActionExpiration :: !(HKD f EpochNo)
-  -- ^ Gov action expiration in number of Epochs
+  , cppGovActionLifetime :: !(HKD f EpochNo)
+  -- ^ Gov action lifetime in number of Epochs
   , cppGovActionDeposit :: !(HKD f Coin)
   -- ^ The amount of the Gov Action deposit
   , cppDRepDeposit :: !(HKD f Coin)
@@ -188,7 +188,7 @@ data UpgradeConwayPParams f = UpgradeConwayPParams
   , ucppDRepVotingThresholds :: !(HKD f DRepVotingThresholds)
   , ucppMinCommitteeSize :: !(HKD f Natural)
   , ucppCommitteeTermLimit :: !(HKD f Natural)
-  , ucppGovActionExpiration :: !(HKD f EpochNo)
+  , ucppGovActionLifetime :: !(HKD f EpochNo)
   , ucppGovActionDeposit :: !(HKD f Coin)
   , ucppDRepDeposit :: !(HKD f Coin)
   , ucppDRepActivity :: !(HKD f EpochNo)
@@ -224,7 +224,7 @@ instance Default (UpgradeConwayPParams Identity) where
       , ucppDRepVotingThresholds = def
       , ucppMinCommitteeSize = 0
       , ucppCommitteeTermLimit = 0
-      , ucppGovActionExpiration = EpochNo 0
+      , ucppGovActionLifetime = EpochNo 0
       , ucppGovActionDeposit = Coin 0
       , ucppDRepDeposit = Coin 0
       , ucppDRepActivity = EpochNo 0
@@ -237,7 +237,7 @@ instance Default (UpgradeConwayPParams StrictMaybe) where
       , ucppDRepVotingThresholds = SNothing
       , ucppMinCommitteeSize = SNothing
       , ucppCommitteeTermLimit = SNothing
-      , ucppGovActionExpiration = SNothing
+      , ucppGovActionLifetime = SNothing
       , ucppGovActionDeposit = SNothing
       , ucppDRepDeposit = SNothing
       , ucppDRepActivity = SNothing
@@ -251,7 +251,7 @@ instance EncCBOR (UpgradeConwayPParams Identity) where
         !> To ucppDRepVotingThresholds
         !> To ucppMinCommitteeSize
         !> To ucppCommitteeTermLimit
-        !> To ucppGovActionExpiration
+        !> To ucppGovActionLifetime
         !> To ucppGovActionDeposit
         !> To ucppDRepDeposit
         !> To ucppDRepActivity
@@ -350,7 +350,7 @@ instance Crypto c => ConwayEraPParams (ConwayEra c) where
       <*> pGroup GovernanceGroup cppDRepVotingThresholds
       <*> pGroup GovernanceGroup cppMinCommitteeSize
       <*> pGroup GovernanceGroup cppCommitteeTermLimit
-      <*> pGroup GovernanceGroup cppGovActionExpiration
+      <*> pGroup GovernanceGroup cppGovActionLifetime
       <*> pGroup GovernanceGroup cppGovActionDeposit
       <*> pGroup GovernanceGroup cppDRepDeposit
       <*> pGroup GovernanceGroup cppDRepActivity
@@ -364,7 +364,7 @@ instance Crypto c => ConwayEraPParams (ConwayEra c) where
       , isValid (/= 0) ppuMaxValSizeL
       , isValid (/= 0) ppuCollateralPercentageL
       , isValid (/= 0) ppuCommitteeTermLimitL
-      , isValid (/= EpochNo 0) ppuGovActionExpirationL
+      , isValid (/= EpochNo 0) ppuGovActionLifetimeL
       , -- Coins
         isValid (/= zero) ppuPoolDepositL
       , isValid (/= zero) ppuGovActionDepositL
@@ -383,7 +383,7 @@ instance Crypto c => ConwayEraPParams (ConwayEra c) where
   hkdDRepVotingThresholdsL = lens cppDRepVotingThresholds (\pp x -> pp {cppDRepVotingThresholds = x})
   hkdMinCommitteeSizeL = lens cppMinCommitteeSize (\pp x -> pp {cppMinCommitteeSize = x})
   hkdCommitteeTermLimitL = lens cppCommitteeTermLimit (\pp x -> pp {cppCommitteeTermLimit = x})
-  hkdGovActionExpirationL = lens cppGovActionExpiration (\pp x -> pp {cppGovActionExpiration = x})
+  hkdGovActionLifetimeL = lens cppGovActionLifetime (\pp x -> pp {cppGovActionLifetime = x})
   hkdGovActionDepositL = lens cppGovActionDeposit (\pp x -> pp {cppGovActionDeposit = x})
   hkdDRepDepositL = lens cppDRepDeposit (\pp x -> pp {cppDRepDeposit = x})
   hkdDRepActivityL = lens cppDRepActivity (\pp x -> pp {cppDRepActivity = x})
@@ -419,7 +419,7 @@ instance Era era => EncCBOR (ConwayPParams Identity era) where
         !> To cppDRepVotingThresholds
         !> To cppMinCommitteeSize
         !> To cppCommitteeTermLimit
-        !> To cppGovActionExpiration
+        !> To cppGovActionLifetime
         !> To cppGovActionDeposit
         !> To cppDRepDeposit
         !> To cppDRepActivity
@@ -509,7 +509,7 @@ emptyConwayPParams =
     , cppDRepVotingThresholds = def
     , cppMinCommitteeSize = 0
     , cppCommitteeTermLimit = 0
-    , cppGovActionExpiration = EpochNo 0
+    , cppGovActionLifetime = EpochNo 0
     , cppGovActionDeposit = Coin 0
     , cppDRepDeposit = Coin 0
     , cppDRepActivity = EpochNo 0
@@ -545,7 +545,7 @@ emptyConwayPParamsUpdate =
     , cppDRepVotingThresholds = SNothing
     , cppMinCommitteeSize = SNothing
     , cppCommitteeTermLimit = SNothing
-    , cppGovActionExpiration = SNothing
+    , cppGovActionLifetime = SNothing
     , cppGovActionDeposit = SNothing
     , cppDRepDeposit = SNothing
     , cppDRepActivity = SNothing
@@ -583,7 +583,7 @@ encodePParamsUpdate ppup =
     !> omitStrictMaybe 26 (cppDRepVotingThresholds ppup) encCBOR
     !> omitStrictMaybe 27 (cppMinCommitteeSize ppup) encCBOR
     !> omitStrictMaybe 28 (cppCommitteeTermLimit ppup) encCBOR
-    !> omitStrictMaybe 29 (cppGovActionExpiration ppup) encCBOR
+    !> omitStrictMaybe 29 (cppGovActionLifetime ppup) encCBOR
     !> omitStrictMaybe 30 (cppGovActionDeposit ppup) encCBOR
     !> omitStrictMaybe 31 (cppDRepDeposit ppup) encCBOR
     !> omitStrictMaybe 32 (cppDRepActivity ppup) encCBOR
@@ -627,7 +627,7 @@ updateField = \case
   26 -> field (\x up -> up {cppDRepVotingThresholds = SJust x}) From
   27 -> field (\x up -> up {cppMinCommitteeSize = SJust x}) From
   28 -> field (\x up -> up {cppCommitteeTermLimit = SJust x}) From
-  29 -> field (\x up -> up {cppGovActionExpiration = SJust x}) From
+  29 -> field (\x up -> up {cppGovActionLifetime = SJust x}) From
   30 -> field (\x up -> up {cppGovActionDeposit = SJust x}) From
   31 -> field (\x up -> up {cppDRepDeposit = SJust x}) From
   32 -> field (\x up -> up {cppDRepActivity = SJust x}) From
@@ -680,7 +680,7 @@ conwayUpgradePParamsHKDPairs px pp =
   , ("dRepVotingThresholds", hkdMap px (toJSON @DRepVotingThresholds) (pp ^. hkdDRepVotingThresholdsL @era @f))
   , ("minCommitteeSize", hkdMap px (toJSON @Natural) (pp ^. hkdMinCommitteeSizeL @era @f))
   , ("committeeTermLimit", hkdMap px (toJSON @Natural) (pp ^. hkdCommitteeTermLimitL @era @f))
-  , ("govActionExpiration", hkdMap px (toJSON @EpochNo) (pp ^. hkdGovActionExpirationL @era @f))
+  , ("govActionLifetime", hkdMap px (toJSON @EpochNo) (pp ^. hkdGovActionLifetimeL @era @f))
   , ("govActionDeposit", hkdMap px (toJSON @Coin) (pp ^. hkdGovActionDepositL @era @f))
   , ("dRepDeposit", hkdMap px (toJSON @Coin) (pp ^. hkdDRepDepositL @era @f))
   , ("dRepActivity", hkdMap px (toJSON @EpochNo) (pp ^. hkdDRepActivityL @era @f))
@@ -700,7 +700,7 @@ upgradeConwayPParamsHKDPairs UpgradeConwayPParams {..} =
   , ("dRepVotingThresholds", (toJSON @DRepVotingThresholds) ucppDRepVotingThresholds)
   , ("minCommitteeSize", (toJSON @Natural) ucppMinCommitteeSize)
   , ("committeeTermLimit", (toJSON @Natural) ucppCommitteeTermLimit)
-  , ("govActionExpiration", (toJSON @EpochNo) ucppGovActionExpiration)
+  , ("govActionLifetime", (toJSON @EpochNo) ucppGovActionLifetime)
   , ("govActionDeposit", (toJSON @Coin) ucppGovActionDeposit)
   , ("dRepDeposit", (toJSON @Coin) ucppDRepDeposit)
   , ("dRepActivity", (toJSON @EpochNo) ucppDRepActivity)
@@ -718,7 +718,7 @@ instance FromJSON (UpgradeConwayPParams Identity) where
         <*> o .: "dRepVotingThresholds"
         <*> o .: "minCommitteeSize"
         <*> o .: "committeeTermLimit"
-        <*> o .: "govActionExpiration"
+        <*> o .: "govActionLifetime"
         <*> o .: "govActionDeposit"
         <*> o .: "dRepDeposit"
         <*> o .: "dRepActivity"
@@ -757,7 +757,7 @@ upgradeConwayPParams UpgradeConwayPParams {..} BabbagePParams {..} =
     , cppDRepVotingThresholds = ucppDRepVotingThresholds
     , cppMinCommitteeSize = ucppMinCommitteeSize
     , cppCommitteeTermLimit = ucppCommitteeTermLimit
-    , cppGovActionExpiration = ucppGovActionExpiration
+    , cppGovActionLifetime = ucppGovActionLifetime
     , cppGovActionDeposit = ucppGovActionDeposit
     , cppDRepDeposit = ucppDRepDeposit
     , cppDRepActivity = ucppDRepActivity
@@ -826,7 +826,7 @@ class BabbageEraPParams era => ConwayEraPParams era where
   hkdDRepVotingThresholdsL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f DRepVotingThresholds)
   hkdMinCommitteeSizeL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Natural)
   hkdCommitteeTermLimitL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Natural)
-  hkdGovActionExpirationL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f EpochNo)
+  hkdGovActionLifetimeL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f EpochNo)
   hkdGovActionDepositL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)
   hkdDRepDepositL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f Coin)
   hkdDRepActivityL :: HKDFunctor f => Lens' (PParamsHKD f era) (HKD f EpochNo)
@@ -843,8 +843,8 @@ ppMinCommitteeSizeL = ppLens . hkdMinCommitteeSizeL @era @Identity
 ppCommitteeTermLimitL :: forall era. ConwayEraPParams era => Lens' (PParams era) Natural
 ppCommitteeTermLimitL = ppLens . hkdCommitteeTermLimitL @era @Identity
 
-ppGovActionExpirationL :: forall era. ConwayEraPParams era => Lens' (PParams era) EpochNo
-ppGovActionExpirationL = ppLens . hkdGovActionExpirationL @era @Identity
+ppGovActionLifetimeL :: forall era. ConwayEraPParams era => Lens' (PParams era) EpochNo
+ppGovActionLifetimeL = ppLens . hkdGovActionLifetimeL @era @Identity
 
 ppGovActionDepositL :: forall era. ConwayEraPParams era => Lens' (PParams era) Coin
 ppGovActionDepositL = ppLens . hkdGovActionDepositL @era @Identity
@@ -867,8 +867,8 @@ ppuMinCommitteeSizeL = ppuLens . hkdMinCommitteeSizeL @era @StrictMaybe
 ppuCommitteeTermLimitL :: forall era. ConwayEraPParams era => Lens' (PParamsUpdate era) (StrictMaybe Natural)
 ppuCommitteeTermLimitL = ppuLens . hkdCommitteeTermLimitL @era @StrictMaybe
 
-ppuGovActionExpirationL :: forall era. ConwayEraPParams era => Lens' (PParamsUpdate era) (StrictMaybe EpochNo)
-ppuGovActionExpirationL = ppuLens . hkdGovActionExpirationL @era @StrictMaybe
+ppuGovActionLifetimeL :: forall era. ConwayEraPParams era => Lens' (PParamsUpdate era) (StrictMaybe EpochNo)
+ppuGovActionLifetimeL = ppuLens . hkdGovActionLifetimeL @era @StrictMaybe
 
 ppuGovActionDepositL :: forall era. ConwayEraPParams era => Lens' (PParamsUpdate era) (StrictMaybe Coin)
 ppuGovActionDepositL = ppuLens . hkdGovActionDepositL @era @StrictMaybe